iox-#1196 Remove superfluous static_cast and add detailed explanation to IOX_DISCARD_RESULT

+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/concurrent/loffli.inl
@@ -25,7 +25,7 @@ namespace concurrent
 {
 inline constexpr uint64_t LoFFLi::requiredIndexMemorySize(const uint64_t capacity) noexcept
 {
-    return (static_cast<uint64_t>(capacity) + 1U) * sizeof(LoFFLi::Index_t);
+    return (capacity + 1U) * sizeof(LoFFLi::Index_t);
 }
 
 } // namespace concurrent
